eve: moves kafka output code to new files
